moved quadcopter controller to examples, updated README
