+++ examples/drone/fly_route.py
@@ -1,18 +1,22 @@
 import genesis as gs
 import math
-from genesis.engine.entities.drone_entity import DroneEntity, DronePIDController
+from quadcopter_controller import DronePIDController
+from genesis.engine.entities.drone_entity import DroneEntity
 from genesis.vis.camera import Camera
 
 base_rpm = 14468.429183500699
 min_rpm = 0.9 * base_rpm
 max_rpm = 1.5 * base_rpm
 
+
 def hover(drone: DroneEntity):
     drone.set_propellels_rpm([base_rpm, base_rpm, base_rpm, base_rpm])
 
+
 def clamp(rpm):
     return max(min_rpm, min(int(rpm), max_rpm))
-    
+
+
 def fly_to_point(target, controller: DronePIDController, scene: gs.Scene, cam: Camera):
     drone = controller.drone
     step = 0
@@ -21,10 +25,13 @@ def fly_to_point(target, controller: DronePIDController, scene: gs.Scene, cam: C
     z = target[2] - drone.get_pos()[2]
 
     distance = math.sqrt(x**2 + y**2 + z**2)
-    
-    while (distance > 0.1 and step < 1000):
+
+    while distance > 0.1 and step < 1000:
         [M1, M2, M3, M4] = controller.update(target)
-        M1=clamp(M1); M2=clamp(M2); M3=clamp(M3); M4=clamp(M4)
+        M1 = clamp(M1)
+        M2 = clamp(M2)
+        M3 = clamp(M3)
+        M4 = clamp(M4)
         drone.set_propellels_rpm([M1, M2, M3, M4])
         scene.step()
         cam.render()
@@ -34,65 +41,42 @@ def fly_to_point(target, controller: DronePIDController, scene: gs.Scene, cam: C
         x = drone_pos[0]
         y = drone_pos[1]
         z = drone_pos[2]
-        cam.set_pose(lookat=(x,y,z))
+        cam.set_pose(lookat=(x, y, z))
         x = target[0] - x
         y = target[0] - y
         z = target[0] - z
         distance = math.sqrt(x**2 + y**2 + z**2)
         step += 1
 
+
 def main():
     gs.init(backend=gs.gpu)
-    
-    ##### scene #####
-    scene = gs.Scene(
-        show_viewer=False,
-        sim_options=gs.options.SimOptions(
-            dt=0.01
-        )
 
-    )
+    ##### scene #####
+    scene = gs.Scene(show_viewer=False, sim_options=gs.options.SimOptions(dt=0.01))
 
     ##### entities #####
-    plane = scene.add_entity(
-        morph=gs.morphs.Plane()
-    )
+    plane = scene.add_entity(morph=gs.morphs.Plane())
 
-    drone =  scene.add_entity(
-        morph = gs.morphs.Drone(
-            file="urdf/drones/cf2x.urdf",
-            pos=(0,0,0.2)
-        )
-    )
+    drone = scene.add_entity(morph=gs.morphs.Drone(file="urdf/drones/cf2x.urdf", pos=(0, 0, 0.2)))
 
-    # parameters are tuned such that the 
+    # parameters are tuned such that the
     # drone can fly, not optimized
     pid_params = [
-        [2.,0.,0.],
-        [2.,0.,0.],
-        [2.,0.,0.],
-        [20.,0.,20.],
-        [20.,0.,20.],
-        [25.,0.,20.],
-        [10.,0.,1.],
-        [10.,0.,1.],
-        [2.,0.,0.2],
+        [2.0, 0.0, 0.0],
+        [2.0, 0.0, 0.0],
+        [2.0, 0.0, 0.0],
+        [20.0, 0.0, 20.0],
+        [20.0, 0.0, 20.0],
+        [25.0, 0.0, 20.0],
+        [10.0, 0.0, 1.0],
+        [10.0, 0.0, 1.0],
+        [2.0, 0.0, 0.2],
     ]
 
-    controller = DronePIDController(
-        drone=drone, 
-        dt=0.01,
-        base_rpm=base_rpm,
-        pid_params=pid_params
-        )
+    controller = DronePIDController(drone=drone, dt=0.01, base_rpm=base_rpm, pid_params=pid_params)
 
-    cam = scene.add_camera(
-        pos=(1,1,1),
-        lookat=drone.morph.pos,
-        GUI=False,
-        res = (640,480),
-        fov=30
-    )
+    cam = scene.add_camera(pos=(1, 1, 1), lookat=drone.morph.pos, GUI=False, res=(640, 480), fov=30)
 
     ##### build #####
 
@@ -100,16 +84,13 @@ def main():
 
     cam.start_recording()
 
-    points = [
-        (1,1,2),
-        (-1,2,1),
-        (0,0,0.5)
-    ]
+    points = [(1, 1, 2), (-1, 2, 1), (0, 0, 0.5)]
 
     for point in points:
         fly_to_point(point, controller, scene, cam)
 
     cam.stop_recording(save_to_filename="../../videos/fly_route.mp4")
+
 
 if __name__ == "__main__":
     main()
